+++ nix/mascot-package.nix
@@ -4,10 +4,7 @@ pkgs.stdenvNoCC.mkDerivation {
   pname = "inir-mascot";
   version = "3";
 
-  src = pkgs.fetchurl {
-    url = "https://github.com/snowarch/inir-mascot/releases/download/v3/inir-mascot-pack.tar.gz";
-    hash = "sha256-DCkWHOVa/7N9FlGD+XdVBuyXRnlWf+3Kv3Lp9f9aw5s=";
-  };
+  src = (import ./mascot-pack.nix { inherit pkgs; }).src;
 
   dontUnpack = true;